+++ client/main.rs
@@ -16,12 +16,7 @@ use crate::client::Client;
 #[async_std::main]
 async fn main() -> Result<()> {
     // Get settings from the configuration file and the program defaults.
-    let settings = Settings::new()?;
-    // Immediately save it. This also creates the save file in case it didn't exist yet.
-    if let Err(error) = settings.save() {
-        println!("Failed saving config file.");
-        println!("{:?}", error);
-    }
+    let settings = Settings::new(true)?;
 
     // Parse commandline options.
     let opt = Opt::from_args();